Refactor: move task page to the base route
